requirements.txt
Django==3.0.5
djangorestframework==3.11.0
drf-extensions==0.6.0

